+++ geoextent/lib/content_providers/__init__.py
@@ -16,6 +16,7 @@ from . import GFZ
 from . import Pensoft
 from . import Opara
 from . import BGR
+from . import MendeleyData
 
 __all__ = [
     "providers",
@@ -29,4 +30,5 @@ __all__ = [
     "Pensoft",
     "Opara",
     "BGR",
+    "MendeleyData",
 ]